Use skeleton as ground truth

+++ src/tree_ring_analyzer/split_masks.py
@@ -15,11 +15,15 @@ import random
 if __name__ == '__main__':
     input_path = '/home/khietdang/Documents/khiet/treeRing/input'
     masks_path = '/home/khietdang/Documents/khiet/treeRing/masks'
-    big_dis_path = '/home/khietdang/Documents/khiet/treeRing/big_dis_otherrings'
-    pith_path = '/home/khietdang/Documents/khiet/treeRing/pithWhole'
+    pith_path = None
     tile_path = '/home/khietdang/Documents/khiet/treeRing/tile_big_dis_otherrings'
     seed = 42
-    pithWhole = True
+    pithWhole = False
+    holes = True
+    gaussian = True
+    dilate = 10
+    distance = True
+    skeleton = False
 
     random.seed(seed)
     np.random.seed(seed)
@@ -29,16 +33,15 @@ if __name__ == '__main__':
     train, test = train_test_split(masks_list, test_size=0.2, shuffle=True, random_state=seed)
     train, val = train_test_split(train, test_size=0.2, shuffle=True, random_state=seed)
 
-    # if not os.path.exists(big_dis_path):
-    #     os.makedirs(big_dis_path)
+    if pith_path is not None:
+        if os.path.exists(pith_path):
+            shutil.rmtree(pith_path)
+        createFolder(pith_path)
 
-    # if os.path.exists(pith_path):
-    #     shutil.rmtree(pith_path)
-    # createFolder(pith_path)
-
-    if os.path.exists(tile_path):
-        shutil.rmtree(tile_path)
-    createFolder(tile_path)
+    if tile_path is not None:
+        if os.path.exists(tile_path):
+            shutil.rmtree(tile_path)
+        createFolder(tile_path)
 
     for mask_path in masks_list:
         print(mask_path)
@@ -51,12 +54,12 @@ if __name__ == '__main__':
 
         mask[mask == 255] = 1
 
-        pith, other_rings_dis = splitRingsAndPith(mask)
+        pith, other_rings_dis = splitRingsAndPith(mask, dilate, distance, skeleton)
         thres = np.max(other_rings_dis)
 
         if mask_path in train:
             save_type = 'train'
-            num = 300
+            num = 100
         elif mask_path in test:
             save_type = 'test'
             num = 1
@@ -65,24 +68,34 @@ if __name__ == '__main__':
             num = 1
 
         if mask_path in test or mask_path in val:
-            # savePith(mask_path, pith, image, 0, pith_path, save_type, False, pithWhole)
-            # saveTile(mask_path, other_rings_dis, image, 0, tile_path, save_type, False, thres)
-            saveTileHoles(mask_path, other_rings_dis, image, 0, tile_path, save_type, False, thres)
-            # saveTileHolesGaussian(mask_path, other_rings_dis, image, 0, tile_path, save_type, False, thres)
+            if pith_path is not None:
+                savePith(mask_path, pith, image, 0, pith_path, save_type, False, pithWhole)
+            if tile_path is not None:
+                if not holes:
+                    saveTile(mask_path, other_rings_dis, image, 0, tile_path, save_type, False, thres)
+                elif not gaussian:
+                    saveTileHoles(mask_path, other_rings_dis, image, 0, tile_path, save_type, False, thres)
+                else:
+                    saveTileHolesGaussian(mask_path, other_rings_dis, image, 0, tile_path, save_type, False, thres)
         else:
-            # data = []
-            # for i in range(0, num):
-            #     data.append((mask_path, pith, image, i, pith_path, save_type, True, pithWhole))
+            if pith_path is not None:
+                data = []
+                for i in range(0, num):
+                    data.append((mask_path, pith, image, i, pith_path, save_type, True, pithWhole))
 
-            # with Pool(int(multiprocessing.cpu_count())) as pool:
-            #     pool.starmap(savePith, data)
+                with Pool(int(multiprocessing.cpu_count())) as pool:
+                    pool.starmap(savePith, data)
+            
+            if tile_path is not None:
+                data = []
+                for i in range(0, num):
+                    data.append((mask_path, other_rings_dis, image, i, tile_path, save_type, True, thres))
 
-            data = []
-            for i in range(0, num):
-                data.append((mask_path, other_rings_dis, image, i, tile_path, save_type, True, thres))
-
-            with Pool(int(os.cpu_count() * 0.5)) as pool:
-                # pool.starmap(saveTile, data)
-                pool.starmap(saveTileHoles, data)
-                # pool.starmap(saveTileHolesGaussian, data)
+                with Pool(int(multiprocessing.cpu_count() * 0.5)) as pool:
+                    if not holes:
+                        pool.starmap(saveTile, data)
+                    elif not gaussian:
+                        pool.starmap(saveTileHoles, data)
+                    else:
+                        pool.starmap(saveTileHolesGaussian, data)
